Moved temp folder creation to FileUtils
