+++ src/Dotnet.Script.DependencyModel/Environment/RuntimeHelper.cs
@@ -95,37 +95,6 @@ namespace Dotnet.Script.DependencyModel.Environment
             return runtimeIdentifier;
         }
 
-        public static string CreateTempFolder(string targetDirectory)
-        {
-            if (!Path.IsPathRooted(targetDirectory))
-            {
-                throw new ArgumentOutOfRangeException(nameof(targetDirectory), "Must be a root path");
-            }
-
-            var tempDirectory = Path.GetTempPath();
-            var pathRoot = Path.GetPathRoot(targetDirectory);
-            var targetDirectoryWithoutRoot = targetDirectory.Substring(pathRoot.Length);
-            if (pathRoot.Length > 0 && IsWindows)
-            {
-                var driveLetter = pathRoot.Substring(0, 1);
-                if (driveLetter == "\\")
-                {
-                    targetDirectoryWithoutRoot = targetDirectoryWithoutRoot.TrimStart(new char[] { '\\' });
-                    driveLetter = "UNC";
-                }
-
-                targetDirectoryWithoutRoot = Path.Combine(driveLetter, targetDirectoryWithoutRoot);
-            }
-            var pathToProjectDirectory = Path.Combine(tempDirectory, "scripts", targetDirectoryWithoutRoot);
-
-            if (!Directory.Exists(pathToProjectDirectory))
-            {
-                Directory.CreateDirectory(pathToProjectDirectory);
-            }
-
-            return pathToProjectDirectory;
-        }
-
         internal static bool AppliesToCurrentRuntime(string runtime)
         {            
             return string.IsNullOrWhiteSpace(runtime) || RuntimeMatcher.IsMatch(runtime);
