+++ src/Dotnet.Script.DependencyModel/ProjectSystem/FileUtils.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Dotnet.Script.DependencyModel.Environment;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
@@ -16,6 +17,37 @@ namespace Dotnet.Script.DependencyModel.ProjectSystem
                     return reader.ReadToEnd();
                 }
             }
+        }
+
+        public static string CreateTempFolder(string targetDirectory)
+        {
+            if (!Path.IsPathRooted(targetDirectory))
+            {
+                throw new ArgumentOutOfRangeException(nameof(targetDirectory), "Must be a root path");
+            }
+
+            var tempDirectory = Path.GetTempPath();
+            var pathRoot = Path.GetPathRoot(targetDirectory);
+            var targetDirectoryWithoutRoot = targetDirectory.Substring(pathRoot.Length);
+            if (pathRoot.Length > 0 && ScriptEnvironment.Instance.IsWindows)
+            {
+                var driveLetter = pathRoot.Substring(0, 1);
+                if (driveLetter == "\\")
+                {
+                    targetDirectoryWithoutRoot = targetDirectoryWithoutRoot.TrimStart(new char[] { '\\' });
+                    driveLetter = "UNC";
+                }
+
+                targetDirectoryWithoutRoot = Path.Combine(driveLetter, targetDirectoryWithoutRoot);
+            }
+            var pathToProjectDirectory = Path.Combine(tempDirectory, "scripts", targetDirectoryWithoutRoot);
+
+            if (!Directory.Exists(pathToProjectDirectory))
+            {
+                Directory.CreateDirectory(pathToProjectDirectory);
+            }
+
+            return pathToProjectDirectory;
         }
     }
 }
