+++ src/Dotnet.Script.DependencyModel/ProjectSystem/ScriptProjectProvider.cs
@@ -109,7 +109,7 @@ namespace Dotnet.Script.DependencyModel.ProjectSystem
 
         private static string GetPathToProjectFile(string targetDirectory)
         { 
-            var pathToProjectDirectory = RuntimeHelper.CreateTempFolder(targetDirectory);
+            var pathToProjectDirectory = FileUtils.CreateTempFolder(targetDirectory);
             var pathToProjectFile = Path.Combine(pathToProjectDirectory, "script.csproj");
             return pathToProjectFile;
         }
